feat: refactor visitor table
