+++ core/src/entity/visitor_spots.rs
@@ -8,6 +8,7 @@ pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i32,
     pub visitor_id: i32,
+    pub spot_id: i32,
     pub last_scanned_at: DateTime,
     pub last_picked_at: Option<DateTime>,
     pub created_at: DateTime,
@@ -17,6 +18,14 @@ pub struct Model {
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
 pub enum Relation {
     #[sea_orm(
+        belongs_to = "super::event_spots::Entity",
+        from = "Column::SpotId",
+        to = "super::event_spots::Column::Id",
+        on_update = "Cascade",
+        on_delete = "Cascade"
+    )]
+    EventSpots,
+    #[sea_orm(
         belongs_to = "super::visitors::Entity",
         from = "Column::VisitorId",
         to = "super::visitors::Column::Id",
@@ -24,6 +33,12 @@ pub enum Relation {
         on_delete = "Cascade"
     )]
     Visitors,
+}
+
+impl Related<super::event_spots::Entity> for Entity {
+    fn to() -> RelationDef {
+        Relation::EventSpots.def()
+    }
 }
 
 impl Related<super::visitors::Entity> for Entity {
