+++ migration/src/m20230712_215023_create_visitor_table.rs
@@ -1,7 +1,6 @@
 use sea_orm_migration::prelude::*;
 
 use crate::m20230712_175819_create_event_table::Events;
-use crate::m20230712_213646_create_event_spot_table::EventSpots;
 
 #[derive(DeriveMigrationName)]
 pub struct Migration;
@@ -18,13 +17,10 @@ impl Migration {
             .col(ColumnDef::new(Visitors::RegistrationId).string_len(4096).not_null())
             .col(ColumnDef::new(Visitors::IsUpdated).boolean().not_null().default(false))
             .col(ColumnDef::new(Visitors::IsDownloadable).boolean().not_null().default(false))
-            .col(ColumnDef::new(Visitors::LastScannedAt).date_time())
             .col(ColumnDef::new(Visitors::LastDropedAt).date_time())
-            .col(ColumnDef::new(Visitors::LastScannedSpot).integer())
             .col(ColumnDef::new(Visitors::CreatedAt).date_time().default(SimpleExpr::Keyword(Keyword::CurrentTimestamp)).not_null())
             .col(ColumnDef::new(Visitors::UpdatedAt).date_time())
             .foreign_key(foreign_key!(Visitors::EventId to Events::Id Cascade))
-            .foreign_key(foreign_key!(Visitors::LastScannedSpot to EventSpots::Id SetNull))
             .to_owned();
 
         visitor
@@ -67,9 +63,7 @@ pub enum Visitors {
     RegistrationId,
     IsUpdated,
     IsDownloadable,
-    LastScannedAt,
     LastDropedAt,
-    LastScannedSpot,
     CreatedAt,
     UpdatedAt,
 }
@@ -107,13 +101,10 @@ mod tests {
                 r#""registration_id" varchar(4096) NOT NULL,"#,
                 r#""is_updated" bool NOT NULL DEFAULT FALSE,"#,
                 r#""is_downloadable" bool NOT NULL DEFAULT FALSE,"#,
-                r#""last_scanned_at" timestamp without time zone,"#,
                 r#""last_droped_at" timestamp without time zone,"#,
-                r#""last_scanned_spot" integer,"#,
                 r#""created_at" timestamp without time zone DEFAULT CURRENT_TIMESTAMP NOT NULL,"#,
                 r#""updated_at" timestamp without time zone,"#,
-                r#"FOREIGN KEY ("event_id") REFERENCES "events" ("id") ON DELETE CASCADE ON UPDATE CASCADE,"#,
-                r#"FOREIGN KEY ("last_scanned_spot") REFERENCES "event_spots" ("id") ON DELETE SET NULL ON UPDATE SET NULL"#,
+                r#"FOREIGN KEY ("event_id") REFERENCES "events" ("id") ON DELETE CASCADE ON UPDATE CASCADE"#,
                 r#")"#
             ].join(" ")
         );
